Refactor into modular files
